requirements.txt
dbt-snowflake==1.9.0
snowflake-connector-python>=3.10.1
python-dotenv>=1.0.1
slack_sdk>=3.28.0
sqlfluff>=3.0.7
sqlfluff-templater-dbt>=3.0.7
black>=24.8.0
ruff>=0.6.5
